Final Commit
I forgot to make more commits to making the other functions. this is the final working function. it adds the marks and finds the average of all of the marks and makes a dictionary with them all and the user is able to add students and their marks, remove students and print and exit the program.